chore(cleanup): rider code cleanup + import optimization across solution [skip ci]
Solution-wide pass from Rider's "Optimize Imports + Code Cleanup".

- 671 .cs files: unused-using removal, qualified-type → using-aliased
  shortening, variable renames following the storage rename
  (commit 880a9f13 missed some test files; this finishes the rename
  in tests/NoMercy.Tests.MediaProcessing and tests/NoMercy.Tests.Setup)
- packaging/macos/distribution.xml: indent reformat (4-space → 2-space)
- 2 .csproj files: formatting only

Net -294 lines (2131 added / 2425 removed) — consistent with import
cleanup rather than logic changes. NoMercy.Storage builds with zero
warnings; affected projects compile clean against the new driver
naming end-to-end.

+++ src/NoMercyQueue.Sqlite/SqliteQueueContext.cs
@@ -126,7 +126,7 @@ public class SqliteQueueContext : IQueueContext
             .ToList()
             .Select(e => new QueueJobModel
             {
-                Id = (int)e.Id,
+                Id = e.Id,
                 Priority = e.Priority,
                 Queue = e.Queue,
                 Payload = e.Payload,
